move migrate DB script to CLI
