+++ src/mailadm/cmdline.py
@@ -13,7 +13,7 @@ from click import style
 
 import mailadm
 import mailadm.db
-from .conn import DBError
+from .conn import DBError, UserInfo
 from .mailcow import MailcowError
 import mailadm.util
 
@@ -280,6 +280,37 @@ def web(ctx, debug):
     app.run(debug=debug, host="localhost", port=3691)
 
 
+@click.command()
+@click.pass_context
+def migratedb(ctx):
+    db = get_mailadm_db(ctx)
+    with db.write_transaction() as conn:
+        conn.execute("PRAGMA foreign_keys=on;")
+
+        q = "SELECT addr, date, ttl, token_name from users"
+        users = [UserInfo(*args) for args in conn.execute(q).fetchall()]
+        q = "DROP TABLE users"
+        conn.execute(q)
+
+        conn.execute("""
+                    CREATE TABLE users (
+                        addr TEXT PRIMARY KEY,
+                        date INTEGER,
+                        ttl INTEGER,
+                        token_name TEXT NOT NULL,
+                        FOREIGN KEY (token_name) REFERENCES tokens (name)
+                    )
+                """)
+        for u in users:
+            q = """INSERT INTO users (addr, date, ttl, token_name)
+                           VALUES (?, ?, ?, ?)"""
+            conn.execute(q, (u.addr, u.date, u.ttl, u.token_name))
+
+        q = "DELETE FROM config WHERE name=?"
+        conn.execute(q, ("vmail_user",))
+        conn.execute(q, ("path_virtual_mailboxes",))
+
+
 mailadm_main.add_command(init)
 mailadm_main.add_command(config)
 mailadm_main.add_command(list_tokens)
@@ -292,6 +323,7 @@ mailadm_main.add_command(del_user)
 mailadm_main.add_command(list_users)
 mailadm_main.add_command(prune)
 mailadm_main.add_command(web)
+mailadm_main.add_command(migratedb)
 
 
 if __name__ == "__main__":
